test(input): add jest test for onKeyDown prop
